Prefab moedas e aparecendo na interface

+++ Design de Games - Inovacao/Assets/Scripts/Corrida/LinhaDeChegada.cs
@@ -21,7 +21,7 @@ public class LinhaDeChegada : MonoBehaviour
 
 	[Header ("Variáveis das Moedas")]
 	public Points moedas;
-	public float moedasGanhasNessaFase;
+	public int moedasGanhasNessaFase;
 
     public void Start()
     {
